pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "optmix"
version = "0.1.0"
description = "AI Marketing Agents powered by Marketing Mix Modeling — Agent-as-Code for measurement & strategy"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "OptMix Contributors"},
]
keywords = [
    "marketing-mix-modeling",
    "mmm",
    "ai-agents",
    "marketing-analytics",
    "budget-optimization",
    "bayesian",
    "media-mix",
    "multi-agent",
    "agent-as-code",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "pymc>=5.10",
    "pymc-marketing>=0.7",
    "arviz>=0.17",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "plotly>=5.18",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "rich>=13.0",
    "click>=8.1",
    "jinja2>=3.1",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "ipykernel>=6.27",
    "jupyter>=1.0",
]
connectors = [
    "google-ads>=24.0",
    "facebook-business>=19.0",
    "google-analytics-data>=0.18",
    "google-cloud-bigquery>=3.14",
]
lightweight = [
    "lightweight-mmm>=0.1",
    "jax>=0.4",
    "numpyro>=0.14",
]
dashboard = [
    "streamlit>=1.30",
]
reports = [
    "python-pptx>=0.6",
    "weasyprint>=62.0",
]
llm = [
    "anthropic>=0.25",
    "openai>=1.12",
]
all = [
    "optmix[dev,connectors,lightweight,dashboard,reports,llm]",
]

[project.scripts]
optmix = "optmix.cli.main:cli"

[project.urls]
Homepage = "https://github.com/optmix-ai/optmix"
Documentation = "https://optmix-ai.github.io/optmix"
Repository = "https://github.com/optmix-ai/optmix"
Issues = "https://github.com/optmix-ai/optmix/issues"
Changelog = "https://github.com/optmix-ai/optmix/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["optmix"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=optmix --cov-report=term-missing"
markers = ["slow: marks tests as slow (MCMC-based, minutes to run)"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
